Drop python 3.4 from legacy: use new
* Reduce build jobs and remove cythonize code
* backport typehints from modern code and drop pyi
