+++ requirements.txt
@@ -1,2 +1,3 @@
 six >=1.10.0
+futures>=3.1 ; python_version == "2.7"
 typing >= 3.6 ; python_version < "3.8"
