+++ test/test_gevent_threadpooled.py
@@ -24,19 +24,14 @@ try:
 except ImportError:
     gevent = None
 
-import six
-
 import threaded
 
-if six.PY3:
-    from os import cpu_count
-else:
-    try:
-        from multiprocessing import cpu_count
-    except ImportError:
-        def cpu_count():
-            """Fake CPU count."""
-            return 1
+try:
+    from multiprocessing import cpu_count
+except ImportError:
+    def cpu_count():
+        """Fake CPU count."""
+        return 1
 
 
 @unittest.skipIf(gevent is None, 'No gevent')
