+++ tox.ini
@@ -5,7 +5,7 @@
 
 [tox]
 minversion = 2.0
-envlist = pep8, pep257, py{27,34,35,36,37,py,py3}, pylint, docs, bandit, py{34,35,36,37}-nocov, mypy
+envlist = pep8, pep257, py{27,py}, pylint, docs, bandit, mypy
 skipsdist = True
 skip_missing_interpreters = True
 
@@ -18,13 +18,14 @@ deps =
     sphinx
     pytest
     pytest-cov
+    pytest-html
     pytest-sugar
     py{34,35,36}-nocov: Cython
     -r{toxinidir}/CI_REQUIREMENTS.txt
     py{27,py}: mock
 
 commands =
-    py.test -vv --junitxml=unit_result.xml --cov-config .coveragerc --cov-report html --cov=threaded {posargs:test}
+    py.test -vv --junitxml=unit_result.xml --cov-config .coveragerc --cov-report html --self-contained-html --html=report.html --cov=threaded {posargs:test}
     coverage report --fail-under 80
 
 [testenv:py34-nocov]
@@ -85,6 +86,7 @@ usedevelop = False
 commands = pip install ./ -vvv -U
 
 [testenv:pylint]
+usedevelop = False
 deps =
     pylint < 2.0
     -r{toxinidir}/CI_REQUIREMENTS.txt
@@ -109,6 +111,7 @@ count = True
 max-line-length = 120
 
 [testenv:docs]
+usedevelop = False
 deps =
     sphinx
 commands = python setup.py build_sphinx
@@ -121,6 +124,7 @@ deps =
 commands = python setup.py build_sphinx upload_sphinx
 
 [testenv:bandit]
+usedevelop = False
 deps = bandit
 commands = bandit -r threaded
 
@@ -131,6 +135,7 @@ deps =
 commands = pipdeptree
 
 [testenv:mypy]
+usedevelop = False
 deps =
     mypy>=0.620
     -r{toxinidir}/CI_REQUIREMENTS.txt
